# server: add test for `/debug/sub-optimal-plan` HTTP API