--- a/server/sql_info_fetcher.go
+++ b/server/sql_info_fetcher.go
@@ -15,6 +15,7 @@ package server
 
 import (
 	"archive/zip"
+	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -191,7 +192,8 @@ func (sh *sqlInfoFetcher) zipInfoForSQL(w http.ResponseWriter, r *http.Request) 
 		resultChan := make(chan *explainAnalyzeResult)
 		go sh.getExplainAnalyze(ctx, sql, resultChan)
 		errChan := make(chan error)
-		go sh.catchCPUProfile(reqCtx, pprofTime, zw, errChan)
+		var buf bytes.Buffer
+		go sh.catchCPUProfile(reqCtx, pprofTime, &buf, errChan)
 		select {
 		case result := <-resultChan:
 			timer.Stop()
@@ -215,13 +217,29 @@ func (sh *sqlInfoFetcher) zipInfoForSQL(w http.ResponseWriter, r *http.Request) 
 		case <-timer.C:
 			cancelFunc()
 		}
-		err = <-errChan
+		err = dumpCPUProfile(errChan, &buf, zw)
 		if err != nil {
 			err = sh.writeErrFile(zw, "profile.err.txt", err)
 			terror.Log(err)
 			return
 		}
 	}
+}
+
+func dumpCPUProfile(errChan chan error, buf *bytes.Buffer, zw *zip.Writer) error {
+	err := <-errChan
+	if err != nil {
+		return err
+	}
+	fw, err := zw.Create("profile")
+	if err != nil {
+		return err
+	}
+	_, err = fw.Write(buf.Bytes())
+	if err != nil {
+		return err
+	}
+	return nil
 }
 
 func (sh *sqlInfoFetcher) writeErrFile(zw *zip.Writer, name string, err error) error {
@@ -255,14 +273,8 @@ func (sh *sqlInfoFetcher) getExplainAnalyze(ctx context.Context, sql string, res
 	resultChan <- &explainAnalyzeResult{rows: rows}
 }
 
-func (sh *sqlInfoFetcher) catchCPUProfile(ctx context.Context, sec int, zw *zip.Writer, errChan chan<- error) {
-	// dump profile
-	fw, err := zw.Create("profile")
-	if err != nil {
-		errChan <- err
-		return
-	}
-	if err := pprof.StartCPUProfile(fw); err != nil {
+func (sh *sqlInfoFetcher) catchCPUProfile(ctx context.Context, sec int, buf *bytes.Buffer, errChan chan<- error) {
+	if err := pprof.StartCPUProfile(buf); err != nil {
 		errChan <- err
 		return
 	}