<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Improving unit test coverage of `server` package.

### What is changed and how it works?

add test for `/debug/sub-optimal-plan` HTTP API， This API added by #10717

Before this PR:
```
ok  	github.com/pingcap/tidb/server	25.799s	coverage: 65.1% of statements
```
After this PR：
```
ok  	github.com/pingcap/tidb/server	30.378s	coverage: 68.4% of statements
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Integration test

Code changes

 - None

Side effects

 - Possible performance regression

The test time of `server` package increased from 25s to 30s.

Related changes

 - None

Release note

 - None